extract `useFromLocation` to own file and use
